extract and organize
